requirements.txt
asgiref==3.7.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
Django==4.2
django-oauth2-authcodeflow==1.0.1
ecdsa==0.18.0
idna==3.6
pyasn1==0.5.1
pycparser==2.21
python-jose==3.3.0
requests==2.31.0
rsa==4.9
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
urllib3==2.1.0
